+++ CHANGELOG.md
@@ -69,6 +69,11 @@
   retrieve maps even if they're internally stored as empty lists, without having
   to catch exceptions from `Value.assertMap()`.
 
+## 1.26.12
+
+* Fix a bug where nesting properties beneath a Sass-syntax custom property
+  (written as `#{--foo}: ...`) would crash.
+
 ## 1.26.11
 
 * **Potentially breaking bug fix:** `selector.nest()` now throws an error